backend/requirements.txt
flask
flask-cors
python-dotenv
openai
requests
boto3
SQLAlchemy
stripe
google-genai